IDeaRepository v.0.0.62 WORKING: models, dao, service, tests, c3p0, jetty:embedded, remoteMysql connection, war packaging, some jsps, authorizationByLoginAndPassword for maincontroller, button(logout), auth check interceptor
CHANGES: ajax is working from external files, some ajax methods, static metod for AlexBelch, made simple image uploading, creating idea and setting tags are working correctly, dao fixed, images pagination working correctly

TODO: controllers insides, pages, pagination, ajax functions, security
need to make js function to take tags list for cloud, fix the showfiles
Author: Gvo3d

+++ src/main/java/irepdata/dao/ImageDao.java
@@ -11,6 +11,6 @@ public interface ImageDao {
     public void createImage(Image image);
     public Image getImage(Long id);
     public void updateImage(Long id, boolean publicity);
-    public List<Image> getImages(int pagination);
+    public List<Image> getImages(Long pagination);
     public Long getImageCount();
 }
