+++ src/main/java/irepdata/service/ImageService.java
@@ -11,7 +11,7 @@ public interface ImageService {
         public void createImage(Image image);
         public Image getImage(Long id);
         public void updateImage(Long id, boolean publicity);
-        public List<Image> getImages(int pagination);
+        public List<Image> getImages(Long pagination);
         public Long getImageCount();
 
 }
